+++ packages/test-peer/const/index.ts
@@ -7,3 +7,9 @@ export const TMP_IROHA_INSTALL_DIR = '.iroha';
  * Directory where compiled iroha_cli will be put
  */
 export const TMP_IROHA_DEPLOY_DIR = '.iroha-deploy';
+
+/**
+ * - Name of the executable iroha cli binary
+ * - Name of source crate
+ */
+export const IROHA_CLI_NAME = 'iroha';