+++ packages/test-peer/scripts/prepare-iroha.ts
@@ -2,7 +2,7 @@ import execa from 'execa';
 import { series } from 'gulp';
 import path from 'path';
 import makeDir from 'make-dir';
-import { TMP_IROHA_DEPLOY_DIR, TMP_IROHA_INSTALL_DIR } from '../const';
+import { TMP_IROHA_DEPLOY_DIR, TMP_IROHA_INSTALL_DIR, IROHA_CLI_NAME } from '../const';
 
 const GIT_IROHA_REPO = 'https://github.com/hyperledger/iroha.git';
 const GIT_BRANCH = 'iroha2-dev';
@@ -24,7 +24,7 @@ async function install_iroha_binary() {
             '--branch',
             GIT_BRANCH,
             // and finally - which crate from repo to install?
-            'iroha_cli',
+            IROHA_CLI_NAME,
         ],
         { stdio: 'inherit' },
     );
@@ -35,7 +35,7 @@ async function make_deploy_dir() {
 }
 
 async function copy_binary_into_deploy_dir() {
-    await execa('cp', [path.resolve(IROHA_INSTALL_ROOT_DIR, 'bin/iroha_cli'), IROHA_DEPLOY_DIR]);
+    await execa('cp', [path.resolve(IROHA_INSTALL_ROOT_DIR, 'bin', IROHA_CLI_NAME), IROHA_DEPLOY_DIR]);
 }
 
 export default series(install_iroha_binary, make_deploy_dir, copy_binary_into_deploy_dir);